Proof of Concept: Remove saved filters from the User entity.
Improve the saved filters design at the database level by removing the saved filter from the User entity.
Add a `saved` boolean on the NamedSearchFilterGroup entity.

This proof of concept focuses only on the back end and makes the absolute minimal changes (even if not ideal) on the front end.
This attempts to avoid addressing the other problems by entirely ignoring them.

The remove NamedSearchFilterGroup endpoint is now changed to a DELETE request that accepts an ID.
A full entity is not needed and should not be required.
This is to avoid validating an entity to be deleted.
If the entity has bad data, it does not matter.
Just delete it.
The only important thing here is the ID.

Additional checks are performed such that a non-existent entity is not deleted.
Generate a warning but still return SUCCESS as the desired behavior of removing a non-existent ID should equate to that ID not existing after this operation is complete.

Add an additional check after the delete to ensure that the delete is successful.
If the entity by the given ID still exists after the delete, then return ERROR.
(The deleteById() does not appear to return a status designating if delete failed or not.)

The `getAllSaveFilterCriteria()` is now able to be changed into a single function call.
This function call just directly fetches from the NamedSearchFilterGroup table and fetches by (owned by current user AND is saved, OR has public flag set to TRUE).

The `/save-filter-criteria` end point is simplified to a single save action.
There is no data on the user directly anymore, removing the necessity to maintain an additional table in the DB.
The save can now be directly against the NamedSearchFilterGroup entity's table.
The existing behavior of `/save-filter-criteria` is a "create new" and not an "update existing" design, whose behavior is preserved.

The NamedSearchFilterGroup entity now has a `savedFlag`.
This flag designates what is called a "saved" filter.
This is an alternative database design to having the NamedSearchFilterGroup on each User entity in the DB.