Existing bug can cause bad state where public flag is true and saved flag is false.
There shouldn't be a case where public flag is true and saved is false.
A separate bug is causing this.
The presence of a bug doesn't mean this should not be handled.
The bad state shouldn't be presented, regardless if a bug causes this or not.

Add an additional check to ensure that the is saved is true when is public is true.