Fix bug causing the named search filter group to be incorrectly saved.
The call to `namedSearchFilterGroupRepo.clone()` is resulting in changing the data in the `named_search_filter_group` table.
This is breaking the interface and causing a problem.

While it is a  separate issue from this proof of concept, this commit helps show the problem and identify that this is not a regression but an existing issue.
Note that the Now Filtering By panel is not refreshing.
There is likely more work needed in regards to broadcasting.