Improve reporting of the Sword error messages.
The error message returned is misleading.
Change the error message to report the actual reason for failure so that better debugging can be performed.

This takes a conservative approach and re-uses existing error handling.
This existing error handling still alters the message but is more accurate than the previous behavior.

Move the ServiceDocument declaration outside of the try block so that it can be used to determine if the failure is due to being unable to fetch the service document for any other reason.
I don't know why the original error message is not displayed and so I am maintaining the existing behavior.

This is not a complete solution.
Something, somewhere, is still arbitrarily adding "409" as an error code (which is very misleading because the actual error codes might be 400 or 401).