Make the Embargo Type and Submission Type `None` values configurable.
The behavior of the case of `None` is treated as as synonymous with a NULL value for that field in the database.
I've looked over the different use cases and have found that in some cases `Unknown` and `Unassigned` are used.

Make this `None` filter value customizable rather than hard-coding the opinion of `None` being NULL.

The `app.filter.embargoTypeNone` and `app.filter.submissionTypeNone` may now be changed to something other than `None`.
The default remains set to `None`.